move a functions to utilities
